Remove GameDetails component and move GameOverview as GameDashboard
