--- newIDE/app/src/GameDashboard/GameDetails.js
@@ -1,723 +1,0 @@
-// @flow
-import { Trans, t } from '@lingui/macro';
-import { I18n } from '@lingui/react';
-import type { Limits } from '../Utils/GDevelopServices/Usage';
-import { type I18n as I18nType } from '@lingui/core';
-import * as React from 'react';
-import FlatButton from '../UI/FlatButton';
-import { Column, Line, Spacer } from '../UI/Grid';
-import {
-  type Game,
-  updateGame,
-  deleteGame,
-  getPublicGame,
-  setGameUserAcls,
-  setGameSlug,
-  getAclsFromUserIds,
-  getCategoryName,
-  getGameMainImageUrl,
-} from '../Utils/GDevelopServices/Game';
-import { type Build } from '../Utils/GDevelopServices/Build';
-import { type TabOptions } from '../UI/Tabs';
-import { ColumnStackLayout, ResponsiveLineStackLayout } from '../UI/Layout';
-import Text from '../UI/Text';
-import AuthenticatedUserContext from '../Profile/AuthenticatedUserContext';
-import PlaceholderError from '../UI/PlaceholderError';
-import Chip from '@material-ui/core/Chip';
-import Builds from '../ExportAndShare/Builds';
-import AlertMessage from '../UI/AlertMessage';
-import RaisedButton from '../UI/RaisedButton';
-import { type PublicGame } from '../Utils/GDevelopServices/Game';
-import PlaceholderLoader from '../UI/PlaceholderLoader';
-import {
-  PublicGamePropertiesDialog,
-  type PublicGameAndProjectEditableProperties,
-} from './PublicGamePropertiesDialog';
-import TextField from '../UI/TextField';
-import KeyboardIcon from '@material-ui/icons/Keyboard';
-import SportsEsportsIcon from '@material-ui/icons/SportsEsports';
-import SmartphoneIcon from '@material-ui/icons/Smartphone';
-import Crown from '../UI/CustomSvgIcons/Crown';
-import { showErrorBox } from '../UI/Messages/MessageBox';
-import LeaderboardAdmin from './LeaderboardAdmin';
-import { GameAnalyticsPanel } from './GameAnalyticsPanel';
-import GameFeedback from './Feedbacks/GameFeedback';
-import GameMonetization from './Monetization/GameMonetization';
-import RouterContext from '../MainFrame/RouterContext';
-import { sendGameDetailsOpened } from '../Utils/Analytics/EventSender';
-import useAlertDialog from '../UI/Alert/useAlertDialog';
-import { extractGDevelopApiErrorStatusAndCode } from '../Utils/GDevelopServices/Errors';
-import CreditsStatusBanner from '../Credits/CreditsStatusBanner';
-import MarketingPlans from '../MarketingPlans/MarketingPlans';
-import MultiplayerAdmin from './MultiplayerAdmin';
-import { GameThumbnail } from './GameThumbnail';
-import { getBuilds } from '../Utils/GDevelopServices/Build';
-
-export type GameDetailsTab =
-  | 'details'
-  | 'builds'
-  | 'feedback'
-  | 'analytics'
-  | 'multiplayer'
-  | 'leaderboards'
-  | 'marketing';
-
-export const getGameDetailsTabs = (
-  userLimits?: ?Limits
-): TabOptions<GameDetailsTab> => {
-  const hidePremiumProducts =
-    !!userLimits &&
-    !!userLimits.capabilities.classrooms &&
-    userLimits.capabilities.classrooms.hidePremiumProducts;
-
-  return [
-    {
-      value: 'details',
-      label: <Trans>Details</Trans>,
-    },
-    {
-      value: 'builds',
-      label: <Trans>Builds</Trans>,
-    },
-    {
-      value: 'feedback',
-      label: <Trans>Feedback</Trans>,
-    },
-    {
-      value: 'analytics',
-      label: <Trans>Analytics</Trans>,
-    },
-    {
-      value: 'multiplayer',
-      label: <Trans>Multiplayer</Trans>,
-    },
-    {
-      value: 'leaderboards',
-      label: <Trans>Leaderboards</Trans>,
-    },
-    hidePremiumProducts
-      ? null
-      : {
-          value: 'marketing',
-          label: <Trans>Marketing & Ads</Trans>,
-        },
-  ].filter(Boolean);
-};
-
-type Props = {|
-  game: Game,
-  project: ?gdProject,
-  onGameUpdated: (game: Game) => void,
-  onGameDeleted: () => void,
-  onLoading: boolean => void,
-  currentTab: GameDetailsTab,
-  setCurrentTab: GameDetailsTab => void,
-  analyticsSource: 'profile' | 'homepage' | 'projectManager',
-|};
-
-const GameDetails = ({
-  game,
-  project,
-  onGameUpdated,
-  onGameDeleted,
-  onLoading,
-  currentTab,
-  setCurrentTab,
-  analyticsSource,
-}: Props) => {
-  const { routeArguments, removeRouteArguments } = React.useContext(
-    RouterContext
-  );
-  const authenticatedUser = React.useContext(AuthenticatedUserContext);
-  const { getAuthorizationHeader, profile, limits } = authenticatedUser;
-  const [builds, setBuilds] = React.useState<?(Build[])>(null);
-  const [
-    gameUnregisterErrorText,
-    setGameUnregisterErrorText,
-  ] = React.useState<?string>(null);
-  const [isGameUpdating, setIsGameUpdating] = React.useState(false);
-  const { showConfirmation, showAlert } = useAlertDialog();
-  const [publicGame, setPublicGame] = React.useState<?PublicGame>(null);
-  const [publicGameError, setPublicGameError] = React.useState<?Error>(null);
-  const [
-    isPublicGamePropertiesDialogOpen,
-    setIsPublicGamePropertiesDialogOpen,
-  ] = React.useState(false);
-
-  const userId = profile ? profile.id : null;
-
-  // If a game dashboard tab is specified, switch to it.
-  React.useEffect(
-    () => {
-      if (routeArguments['games-dashboard-tab']) {
-        // Ensure that the tab is valid.
-        const gameDetailsTab = getGameDetailsTabs(limits).find(
-          gameDetailsTab =>
-            gameDetailsTab.value === routeArguments['games-dashboard-tab']
-        );
-        if (gameDetailsTab) setCurrentTab(gameDetailsTab.value);
-        // Cleanup once open, to ensure it is not opened again.
-        removeRouteArguments(['games-dashboard-tab']);
-      }
-    },
-    [routeArguments, removeRouteArguments, setCurrentTab, limits]
-  );
-
-  const loadPublicGame = React.useCallback(
-    async () => {
-      setPublicGameError(null);
-      try {
-        const publicGameResponse = await getPublicGame(game.id);
-        setPublicGame(publicGameResponse);
-      } catch (err) {
-        console.error(`Unable to load the game:`, err);
-        setPublicGameError(err);
-      }
-    },
-    [game]
-  );
-
-  const loadBuilds = React.useCallback(
-    async () => {
-      if (!userId) return;
-      const builds = await getBuilds(getAuthorizationHeader, userId, game.id);
-      setBuilds(builds);
-    },
-    [getAuthorizationHeader, userId, game.id]
-  );
-
-  React.useEffect(
-    () => {
-      loadPublicGame();
-      loadBuilds();
-    },
-    [loadPublicGame, loadBuilds]
-  );
-
-  React.useEffect(
-    () => {
-      sendGameDetailsOpened({ from: analyticsSource });
-    },
-    [analyticsSource]
-  );
-
-  const onUpdateGame = async (
-    i18n: I18nType,
-    properties: PublicGameAndProjectEditableProperties
-  ) => {
-    if (!profile || !publicGame) return false;
-    const { ownerIds, userSlug, gameSlug, authorIds } = properties;
-
-    if (!ownerIds || !ownerIds.length) {
-      await showAlert({
-        title: t`Select an owner`,
-        message: t`You must select at least one user to be the owner of the game.`,
-      });
-      return false;
-    }
-
-    if (!authorIds || !authorIds.length) {
-      await showAlert({
-        title: t`Select an author`,
-        message: t`You must select at least one user to be the author of the game.`,
-      });
-      return false;
-    }
-
-    try {
-      setIsGameUpdating(true);
-      const updatedGame = await updateGame(
-        getAuthorizationHeader,
-        profile.id,
-        publicGame.id,
-        {
-          gameName: properties.gameName || 'Untitled game',
-          authorName: properties.authorName || 'Unspecified publisher',
-          categories: properties.categories,
-          publicWebBuildId: properties.publicWebBuildId,
-          description: properties.description,
-          playWithKeyboard: properties.playWithKeyboard,
-          playWithGamepad: properties.playWithGamepad,
-          playWithMobile: properties.playWithMobile,
-          orientation: properties.orientation,
-          thumbnailUrl: properties.thumbnailUrl,
-          screenshotUrls: properties.screenshotUrls,
-          discoverable: properties.discoverable,
-          acceptsBuildComments: properties.acceptsBuildComments,
-          acceptsGameComments: properties.acceptsGameComments,
-          displayAdsOnGamePage: properties.displayAdsOnGamePage,
-        }
-      );
-
-      if (userSlug && gameSlug && userSlug === profile.username) {
-        try {
-          await setGameSlug(
-            getAuthorizationHeader,
-            profile.id,
-            publicGame.id,
-            userSlug,
-            gameSlug
-          );
-        } catch (error) {
-          console.error(
-            'Unable to update the game slug:',
-            error.response || error.message
-          );
-          showErrorBox({
-            message:
-              i18n._(
-                t`Unable to update the game slug. A slug must be 6 to 30 characters long and only contains letters, digits or dashes.`
-              ) +
-              ' ' +
-              i18n._(t`Verify your internet connection or try again later.`),
-            rawError: error,
-            errorId: 'game-slug-update-error',
-          });
-          return false;
-        }
-      }
-      try {
-        const authorAcls = getAclsFromUserIds(authorIds);
-        const ownerAcls = getAclsFromUserIds(ownerIds);
-        await setGameUserAcls(
-          getAuthorizationHeader,
-          profile.id,
-          publicGame.id,
-          {
-            ownership: ownerAcls,
-            author: authorAcls,
-          }
-        );
-      } catch (error) {
-        console.error(
-          'Unable to update the game owners or authors:',
-          error.response || error.message
-        );
-        showErrorBox({
-          message:
-            i18n._(
-              t`Unable to update the game owners or authors. Have you removed yourself from the owners?`
-            ) +
-            ' ' +
-            i18n._(t`Verify your internet connection or try again later.`),
-          rawError: error,
-          errorId: 'game-acls-update-error',
-        });
-        return false;
-      }
-      onGameUpdated(updatedGame);
-    } catch (error) {
-      console.error(
-        'Unable to update the game:',
-        error.response || error.message
-      );
-      showErrorBox({
-        message:
-          i18n._(t`Unable to update the game details.`) +
-          ' ' +
-          i18n._(t`Verify your internet connection or try again later.`),
-        rawError: error,
-        errorId: 'game-details-update-error',
-      });
-      return false;
-    } finally {
-      setIsGameUpdating(false);
-    }
-
-    return true;
-  };
-
-  const updateProjectFromGameIfMatching = (
-    properties: PublicGameAndProjectEditableProperties
-  ) => {
-    if (project && project.getProjectUuid() === game.id) {
-      // Get this information from the game object that should have just been updated
-      // (maybe with fallback values).
-      const { gameName, categories, description } = game;
-      // Get those properties from the object as they are not stored on the Game.
-      const { authorIds, authorUsernames } = properties;
-
-      project.setName(gameName);
-      project.setDescription(description || '');
-      project.setPlayableWithKeyboard(game.playWithKeyboard);
-      project.setPlayableWithGamepad(game.playWithGamepad);
-      project.setPlayableWithMobile(game.playWithMobile);
-      project.setOrientation(game.orientation || 'default');
-      if (categories) {
-        const projectCategories = project.getCategories();
-        projectCategories.clear();
-        categories.forEach(category => projectCategories.push_back(category));
-      }
-      if (authorIds) {
-        const projectAuthorIds = project.getAuthorIds();
-        projectAuthorIds.clear();
-        authorIds.forEach(authorId => projectAuthorIds.push_back(authorId));
-      }
-      if (authorUsernames) {
-        const projectAuthorUsernames = project.getAuthorUsernames();
-        projectAuthorUsernames.clear();
-        authorUsernames.forEach(authorUsername =>
-          projectAuthorUsernames.push_back(authorUsername)
-        );
-      }
-    }
-  };
-
-  const unregisterGame = React.useCallback(
-    async (i18n: I18nType) => {
-      if (!profile) return;
-      const { id } = profile;
-      setGameUnregisterErrorText(null);
-      onLoading(true);
-      try {
-        setIsGameUpdating(true);
-        await deleteGame(getAuthorizationHeader, id, game.id);
-        onGameDeleted();
-      } catch (error) {
-        console.error('Unable to delete the game:', error);
-        const extractedStatusAndCode = extractGDevelopApiErrorStatusAndCode(
-          error
-        );
-        if (
-          extractedStatusAndCode &&
-          extractedStatusAndCode.code === 'game-deletion/leaderboards-exist'
-        ) {
-          setGameUnregisterErrorText(
-            i18n._(
-              t`You cannot unregister a game that has active leaderboards. To delete them, go in the Leaderboards tab, and delete them one by one.`
-            )
-          );
-        }
-      } finally {
-        setIsGameUpdating(false);
-        onLoading(false);
-      }
-    },
-    [onLoading, game.id, profile, onGameDeleted, getAuthorizationHeader]
-  );
-
-  const unpublishGame = React.useCallback(
-    async () => {
-      if (!profile) return;
-
-      const { id } = profile;
-      try {
-        setIsGameUpdating(true);
-        const updatedGame = await updateGame(
-          getAuthorizationHeader,
-          id,
-          game.id,
-          {
-            publicWebBuildId: null,
-          }
-        );
-        onGameUpdated(updatedGame);
-      } catch (err) {
-        console.error('Unable to update the game', err);
-      } finally {
-        setIsGameUpdating(false);
-      }
-    },
-    [game, getAuthorizationHeader, profile, onGameUpdated]
-  );
-
-  const onClickUnregister = React.useCallback(
-    async (i18n: I18nType) => {
-      const answer = await showConfirmation({
-        title: t`Unregister game`,
-        message: t`Are you sure you want to unregister this game?${'\n\n'}It will disappear from your games dashboard and you won't get access to analytics, unless you register it again.`,
-      });
-
-      if (!answer) return;
-
-      unregisterGame(i18n);
-    },
-    [unregisterGame, showConfirmation]
-  );
-
-  const onClickUnpublish = React.useCallback(
-    async (i18n: I18nType) => {
-      const answer = await showConfirmation({
-        title: t`Unpublish game`,
-        message: t`Are you sure you want to unpublish this game?${'\n\n'}This will make your gd.games unique game URL not accessible anymore.${'\n\n'}You can decide at any time to publish it again.`,
-      });
-
-      if (!answer) return;
-
-      unpublishGame();
-    },
-    [unpublishGame, showConfirmation]
-  );
-
-  const authorUsernames =
-    publicGame &&
-    publicGame.authors.map(author => author.username).filter(Boolean);
-
-  const ownerUsernames =
-    publicGame &&
-    publicGame.owners.map(owner => owner.username).filter(Boolean);
-
-  const isGameOpenedAsProject =
-    !!project && project.getProjectUuid() === game.id;
-
-  const publicGameThumbnailUrl = React.useMemo(
-    () => (publicGame ? getGameMainImageUrl(publicGame) : null),
-    [publicGame]
-  );
-
-  const webBuilds = builds
-    ? builds.filter(build => build.type === 'web-build')
-    : null;
-
-  const canBePublishedOnGdGames = webBuilds ? webBuilds.length > 0 : false;
-
-  return (
-    <I18n>
-      {({ i18n }) => (
-        <>
-          <Line expand noMargin useFullHeight>
-            {currentTab === 'leaderboards' ? (
-              <LeaderboardAdmin gameId={game.id} onLoading={onLoading} />
-            ) : null}
-            {currentTab === 'multiplayer' ? (
-              <MultiplayerAdmin gameId={game.id} />
-            ) : null}
-            {currentTab === 'details' ? (
-              publicGameError ? (
-                <PlaceholderError onRetry={loadPublicGame}>
-                  <Trans>There was an issue getting the game details.</Trans>{' '}
-                  <Trans>
-                    Verify your internet connection or try again later.
-                  </Trans>
-                </PlaceholderError>
-              ) : !publicGame ? (
-                <PlaceholderLoader />
-              ) : (
-                <ColumnStackLayout expand noMargin>
-                  {!isGameOpenedAsProject && (
-                    <AlertMessage kind="info">
-                      <Trans>
-                        In order to update these details you have to open the
-                        game's project.
-                      </Trans>
-                    </AlertMessage>
-                  )}
-                  <ResponsiveLineStackLayout noMargin>
-                    <Column noMargin alignItems="center">
-                      <GameThumbnail
-                        gameName={publicGame.gameName}
-                        thumbnailUrl={publicGameThumbnailUrl}
-                        gameId={publicGame.id}
-                      />
-                    </Column>
-                    <ColumnStackLayout expand noMargin>
-                      <TextField
-                        value={publicGame.gameName}
-                        readOnly
-                        fullWidth
-                        floatingLabelText={<Trans>Game name</Trans>}
-                        floatingLabelFixed={true}
-                      />
-                      <Line alignItems="center" noMargin>
-                        <Line
-                          expand
-                          justifyContent="flex-start"
-                          alignItems="center"
-                          noMargin
-                        >
-                          {authorUsernames && (
-                            <>
-                              <Text>
-                                <Trans>Authors:</Trans>
-                              </Text>
-                              <Line noMargin>
-                                {authorUsernames.map((username, index) => (
-                                  <React.Fragment key={username}>
-                                    <Spacer />
-                                    <Chip
-                                      size="small"
-                                      icon={
-                                        ownerUsernames &&
-                                        ownerUsernames.includes(username) ? (
-                                          <Crown />
-                                        ) : (
-                                          undefined
-                                        )
-                                      }
-                                      className="notranslate"
-                                      label={username}
-                                      color={
-                                        index === 0 ? 'primary' : 'default'
-                                      }
-                                    />
-                                  </React.Fragment>
-                                ))}
-                              </Line>
-                            </>
-                          )}
-                        </Line>
-                        <Line expand justifyContent="flex-end" noMargin>
-                          <Text>
-                            <Trans>
-                              Created on {i18n.date(game.createdAt * 1000)}
-                            </Trans>
-                          </Text>
-                        </Line>
-                      </Line>
-                      {(publicGame.playWithKeyboard ||
-                        publicGame.playWithGamepad ||
-                        publicGame.playWithMobile ||
-                        publicGame.categories) && (
-                        <Line alignItems="center" noMargin>
-                          <Line
-                            expand
-                            justifyContent="flex-start"
-                            alignItems="center"
-                            noMargin
-                          >
-                            {publicGame.categories &&
-                              !!publicGame.categories.length && (
-                                <>
-                                  <Text>
-                                    <Trans>Genres:</Trans>
-                                  </Text>
-                                  <Line noMargin>
-                                    {publicGame.categories.map(
-                                      (category, index) => (
-                                        <React.Fragment key={category}>
-                                          <Spacer />
-                                          <Chip
-                                            size="small"
-                                            label={getCategoryName(
-                                              category,
-                                              i18n
-                                            )}
-                                            color={
-                                              index === 0
-                                                ? 'primary'
-                                                : 'default'
-                                            }
-                                          />
-                                        </React.Fragment>
-                                      )
-                                    )}
-                                  </Line>
-                                </>
-                              )}
-                          </Line>
-                          <Line justifyContent="flex-end" noMargin>
-                            {publicGame.playWithKeyboard && <KeyboardIcon />}
-                            {publicGame.playWithGamepad && (
-                              <SportsEsportsIcon />
-                            )}
-                            {publicGame.playWithMobile && <SmartphoneIcon />}
-                          </Line>
-                        </Line>
-                      )}
-                    </ColumnStackLayout>
-                  </ResponsiveLineStackLayout>
-                  <TextField
-                    value={publicGame.description || ''}
-                    readOnly
-                    fullWidth
-                    floatingLabelText={<Trans>Game description</Trans>}
-                    floatingLabelFixed={true}
-                    translatableHintText={t`No description set.`}
-                    multiline
-                    rows={5}
-                  />
-                  <TextField
-                    value={
-                      publicGame.orientation === 'default'
-                        ? i18n._('Platform default')
-                        : publicGame.orientation === 'landscape'
-                        ? i18n._('Landscape')
-                        : i18n._('Portrait')
-                    }
-                    readOnly
-                    fullWidth
-                    floatingLabelText={
-                      <Trans>Device orientation (for mobile)</Trans>
-                    }
-                    floatingLabelFixed={true}
-                  />
-                  <ResponsiveLineStackLayout noMargin justifyContent="flex-end">
-                    <FlatButton
-                      onClick={() => onClickUnregister(i18n)}
-                      label={<Trans>Unregister this game</Trans>}
-                      disabled={isGameUpdating}
-                    />
-                    {publicGame.publicWebBuildId && (
-                      <RaisedButton
-                        onClick={() => onClickUnpublish(i18n)}
-                        label={<Trans>Unpublish from gd.games</Trans>}
-                        disabled={isGameUpdating}
-                      />
-                    )}
-                    <RaisedButton
-                      primary
-                      onClick={() => setIsPublicGamePropertiesDialogOpen(true)}
-                      label={<Trans>Edit game details</Trans>}
-                      disabled={!isGameOpenedAsProject || isGameUpdating}
-                    />
-                  </ResponsiveLineStackLayout>
-                  {gameUnregisterErrorText ? (
-                    <PlaceholderError>
-                      {gameUnregisterErrorText}
-                    </PlaceholderError>
-                  ) : null}
-                </ColumnStackLayout>
-              )
-            ) : null}
-            {currentTab === 'builds' ? (
-              <Builds
-                game={game}
-                authenticatedUser={authenticatedUser}
-                onGameUpdated={onGameUpdated}
-              />
-            ) : null}
-            {currentTab === 'analytics' ? (
-              <GameAnalyticsPanel game={game} />
-            ) : null}
-            {currentTab === 'feedback' ? (
-              <GameFeedback
-                i18n={i18n}
-                authenticatedUser={authenticatedUser}
-                game={game}
-              />
-            ) : null}
-            {currentTab === 'marketing' ? (
-              <ColumnStackLayout noMargin expand>
-                <CreditsStatusBanner displayPurchaseAction />
-                <Text size="sub-title">
-                  <Trans>Marketing Campaigns</Trans>
-                </Text>
-                <MarketingPlans game={game} />
-                <GameMonetization game={game} onGameUpdated={onGameUpdated} />
-              </ColumnStackLayout>
-            ) : null}
-          </Line>
-          {publicGame && project && isPublicGamePropertiesDialogOpen && (
-            <PublicGamePropertiesDialog
-              publicGame={publicGame}
-              canBePublishedOnGdGames={canBePublishedOnGdGames}
-              onApply={async properties => {
-                const isGameUpdated = await onUpdateGame(i18n, properties);
-                if (isGameUpdated) {
-                  updateProjectFromGameIfMatching(properties);
-                }
-              }}
-              onClose={() => setIsPublicGamePropertiesDialogOpen(false)}
-              isLoading={isGameUpdating}
-              i18n={i18n}
-              onUpdatingGame={setIsGameUpdating}
-              onGameUpdated={onGameUpdated}
-            />
-          )}
-        </>
-      )}
-    </I18n>
-  );
-};
-
-export default GameDetails;
