+++ newIDE/app/src/GameDashboard/GameLinkAndShareIcons.js
