+++ newIDE/app/src/GameDashboard/Widgets/AnalyticsWidget.js
@@ -4,19 +4,19 @@ import * as React from 'react';
 import { I18n } from '@lingui/react';
 import { Trans } from '@lingui/macro';
 import DashboardWidget from './DashboardWidget';
-import FlatButton from '../UI/FlatButton';
-import ArrowRight from '../UI/CustomSvgIcons/ArrowRight';
-import { ColumnStackLayout, ResponsiveLineStackLayout } from '../UI/Layout';
-import { Column, Line, Spacer } from '../UI/Grid';
-import Text from '../UI/Text';
-import { type Game } from '../Utils/GDevelopServices/Game';
-import { SessionsChart } from '../GameDashboard/GameAnalyticsCharts';
-import { type GameMetrics } from '../Utils/GDevelopServices/Analytics';
-import { buildLastWeekChartData } from '../GameDashboard/GameAnalyticsEvaluator';
-import RaisedButton from '../UI/RaisedButton';
-import Coin from '../Credits/Icons/Coin';
-import MarketingPlansDialog from '../MarketingPlans/MarketingPlansDialog';
-import GameLinkAndShareIcons from './GameLinkAndShareIcons';
+import FlatButton from '../../UI/FlatButton';
+import ArrowRight from '../../UI/CustomSvgIcons/ArrowRight';
+import { ColumnStackLayout, ResponsiveLineStackLayout } from '../../UI/Layout';
+import { Column, Line, Spacer } from '../../UI/Grid';
+import Text from '../../UI/Text';
+import { type Game } from '../../Utils/GDevelopServices/Game';
+import { SessionsChart } from '../GameAnalyticsCharts';
+import { type GameMetrics } from '../../Utils/GDevelopServices/Analytics';
+import { buildLastWeekChartData } from '../GameAnalyticsEvaluator';
+import RaisedButton from '../../UI/RaisedButton';
+import Coin from '../../Credits/Icons/Coin';
+import MarketingPlansDialog from '../../MarketingPlans/MarketingPlansDialog';
+import GameLinkAndShareIcons from '../GameLinkAndShareIcons';
 
 const styles = { loadingSpace: { height: 100 } };
