+++ newIDE/app/src/GameDashboard/Widgets/BuildsWidget.js
@@ -2,15 +2,15 @@
 
 import * as React from 'react';
 import { Trans } from '@lingui/macro';
-import { type Build } from '../Utils/GDevelopServices/Build';
+import { type Build } from '../../Utils/GDevelopServices/Build';
 import DashboardWidget from './DashboardWidget';
-import FlatButton from '../UI/FlatButton';
-import ArrowRight from '../UI/CustomSvgIcons/ArrowRight';
-import { Line } from '../UI/Grid';
-import Text from '../UI/Text';
-import GDevelopThemeContext from '../UI/Theme/GDevelopThemeContext';
-import NotificationDot from './NotificationDot';
-import { ColumnStackLayout } from '../UI/Layout';
+import FlatButton from '../../UI/FlatButton';
+import ArrowRight from '../../UI/CustomSvgIcons/ArrowRight';
+import { Line } from '../../UI/Grid';
+import Text from '../../UI/Text';
+import GDevelopThemeContext from '../../UI/Theme/GDevelopThemeContext';
+import { ColumnStackLayout } from '../../UI/Layout';
+import NotificationDot from '../NotificationDot';
 
 const styles = {
   ongoingExportsContainer: {
